SPARKC-387 + SPARKC-402: Backwards Compatibility Fixes 2.1
SPARKC-387: ColumnType Should be ProtocolVersion Aware

The Connector currently always assumes it is connecting to the newest
version of C*. This leads to some incorrect decisions when checking
against ColumnTypes as PV4 Types are not valid in PV3. To fix this all
of the ColumnType Selection code is now also checks the Protocol Version
and chooses the Correct Type based Protocol Version instead.

In addition, the Integeration tests have been version gated to allow for
accurate testing against previous versions of Cassandra.

SPARKC-402: PartitionKey Indexes Pushdown Incorrect for PV3

In PV3 it is illegal to pushdown an equality predicate based on a
secondary index for a partition key while also specifying inequalities on
the token of that partition key. In PV4 it is legal. The pushdowns have
now been changed to check the PV4 and skip equality pushdowns for
partition keys if PV3 is in use.